requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
attrs==25.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
compressed-rtf==1.0.7
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.28.7
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.3.1
MarkupSafe==3.0.2
narwhals==1.36.0
numpy==2.2.4
olefile==0.47
openai==1.76.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20191110
pillow==11.2.0
playwright==1.51.0
protobuf==5.29.4
pyarrow==19.0.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
PyMuPDF==1.25.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
six==1.12.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
streamlit==1.44.1
tenacity==9.1.2
textract==1.6.5
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.2
